requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.1
requests==2.31.0
bleach>=6.0.0
html5lib>=1.1
python-dotenv==1.0.0